refactor: Refactor bug reporting
