+++ .env.sample
@@ -1,3 +1,6 @@
+# App
+BUG_REPORT_URL=
+
 # AWS
 BUCKET_REGION=
 IDENTITY_POOL_ID=
